refactor: Change Double to Float for performance and consistency in Windkraftanlage and related classes

+++ src/controler/WindkraftanlageRepository.java
@@ -116,16 +116,16 @@ public class WindkraftanlageRepository {
             int objektId = FeldParser.parseGanzzahlSicher(csvFelder[feldIndex++]);
             String name = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String baujahr = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
-            Double gesamtLeistungMW = FeldParser.parseGleitkommaZahlNullbar(
+            Float gesamtLeistungMW = FeldParser.parseGleitkommaZahlNullbar(
                 CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             Integer anzahl = FeldParser.parseGanzzahlNullbar(
                 CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String typ = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String ort = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String landkreis = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
-            Double breitengrad = FeldParser.parseGleitkommaZahlNullbar(
+            Float breitengrad = FeldParser.parseGleitkommaZahlNullbar(
                 CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
-            Double laengengrad = FeldParser.parseGleitkommaZahlNullbar(
+            Float laengengrad = FeldParser.parseGleitkommaZahlNullbar(
                 CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String betreiber = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
             String bemerkungen = FeldParser.leerZuNull(CsvParser.bereinigesFeld(csvFelder[feldIndex++]));
@@ -173,10 +173,10 @@ public class WindkraftanlageRepository {
         for (Windkraftanlage anlage : alleAnlagen) {
             boolean wurdeKorrigiert = false;
             
-            Double alterBreitengrad = anlage.getBreitengrad();
-            Double alterLaengengrad = anlage.getLaengengrad();
-            Double neuerBreitengrad = alterBreitengrad;
-            Double neuerLaengengrad = alterLaengengrad;
+            Float alterBreitengrad = anlage.getBreitengrad();
+            Float alterLaengengrad = anlage.getLaengengrad();
+            Float neuerBreitengrad = alterBreitengrad;
+            Float neuerLaengengrad = alterLaengengrad;
             
             // Korrigiere Breitengrad wenn ungültig
             if (!KoordinatenValidierer.istBreitengradGueltig(alterBreitengrad)) {
