+++ src/util/ZeitStatistiken.java
@@ -13,7 +13,7 @@ import java.util.Map;
  */
 public final class ZeitStatistiken {
 
-    private static final Map<String, Double> ZEITEN = new LinkedHashMap<>();
+    private static final Map<String, Float> ZEITEN = new LinkedHashMap<>();
     private static final Map<String, String> STATS = new LinkedHashMap<>();
 
     private ZeitStatistiken() { }
@@ -27,7 +27,7 @@ public final class ZeitStatistiken {
      * @param name Name der Messung
      * @param millis Zeit in Millisekunden
      */
-    public static void zeichneZeitAuf(String name, double millis) {
+    public static void zeichneZeitAuf(String name, float millis) {
         ZEITEN.put(name, millis);
     }
 
@@ -53,10 +53,10 @@ public final class ZeitStatistiken {
      */
     public static void druckeZusammenfassung() {
         AusgabeManager.gebeAus(Konstanten.ZUSAMMENFASSUNG_UEBERSCHRIFT);
-        double gesamt = 0.0;
+        float gesamt = 0.0f;
         if (!ZEITEN.isEmpty()) {
             AusgabeManager.gebeAus(Konstanten.ZEITEN_UEBERSCHRIFT);
-            for (Map.Entry<String, Double> e : ZEITEN.entrySet()) {
+            for (Map.Entry<String, Float> e : ZEITEN.entrySet()) {
                 AusgabeManager.gebeAusFormat(Konstanten.ZEIT_FORMAT, e.getKey(), e.getValue());
                 gesamt += e.getValue();
             }
